fix: circular dependencies
Closes #873.
