+++ src/utils/addressMap.ts
@@ -1,9 +1,48 @@
-import type { TransferableInput } from '../serializable';
+import type { OutputOwners, TransferableInput } from '../serializable';
 import type { Utxo } from '../serializable/avax/utxo';
 import { Address } from '../serializable/fxs/common';
 import { addressesFromBytes } from './addressesFromBytes';
 import { hexToBuffer } from './buffer';
-import { matchOwners } from './matchOwners';
+
+export type MatchOwnerResult = {
+  sigIndicies: number[];
+  addressMap: AddressMap;
+};
+export const matchOwners = (
+  owners: OutputOwners,
+  inputAddrs: Address[],
+  minIssuanceTime: bigint,
+  sigindices?: number[],
+): MatchOwnerResult | undefined => {
+  if (owners.locktime.value() > minIssuanceTime) {
+    return undefined;
+  }
+
+  const inputAddrSet = new Set(inputAddrs.map((a) => a.toString()));
+  const addressMap = owners.addrs.reduce((agg, addr, i) => {
+    if (
+      agg.size() < owners.threshold.value() &&
+      inputAddrSet.has(addr.value())
+    ) {
+      // only add actual signer addresses if sigindices are known
+      if (sigindices?.length && !sigindices.includes(i)) {
+        return agg;
+      }
+
+      return agg.set(addr, i);
+    }
+    return agg;
+  }, new AddressMap());
+
+  if (addressMap.size() < owners.threshold.value()) {
+    return undefined;
+  }
+
+  return {
+    sigIndicies: Array.from(addressMap.values()),
+    addressMap: addressMap,
+  };
+};
 
 export class AddressMap {
   constructor(initialData: [Address, number][] = []) {
